+++ TM.Digital.Api/Controllers/GameController.cs
@@ -42,15 +42,19 @@ namespace TM.Digital.Api.Controllers
             
             return await GamesService.Instance.JoinSession(gameId, playerName, _hubContext);
         }
-
-
-
-        [Route("addplayer/{gameId}/{playername}/{test}")]
-        public async Task<GameSetup> AddPlayer(Guid gameId, string playername, bool test)
+        [Route("start/{gameId}")]
+        public async Task<bool> Start(Guid gameId)
         {
-            await Task.CompletedTask;
-            return GamesService.Instance.AddPlayer(gameId, playername, test);
+            return await GamesService.Instance.StartGame(gameId, _hubContext);
         }
+
+
+        //[Route("addplayer/{gameId}/{playername}/{test}")]
+        //public async Task<GameSetup> AddPlayer(Guid gameId, string playername, bool test)
+        //{
+        //    await Task.CompletedTask;
+        //    return GamesService.Instance.AddPlayer(gameId, playername, test);
+        //}
 
         [Route("addplayer/setupplayer")]
         public async Task<Player> SetupPlayer(GameSetupSelection selection)
