--- TM.Digital.Client/GameSetupWindow.xaml
@@ -1,102 +1,0 @@
-﻿<Window x:Class="TM.Digital.Client.GameSetupWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:TM.Digital.Client"
-        mc:Ignorable="d"
-        d:DataContext="{d:DesignInstance {x:Type local:GameSetupViewModel}}"
-        Title="GameSetupWindow" d:DesignHeight="450" d:DesignWidth="800">
-    <Grid>
-        <Grid.RowDefinitions>
-            <RowDefinition></RowDefinition>
-            <RowDefinition></RowDefinition>
-            <RowDefinition Height="Auto"></RowDefinition>
-        </Grid.RowDefinitions>
-        <ItemsControl ItemsSource="{Binding CorporationChoices}">
-            <ItemsControl.ItemsPanel>
-                <ItemsPanelTemplate>
-                    <StackPanel Orientation="Horizontal"></StackPanel>
-                </ItemsPanelTemplate>
-            </ItemsControl.ItemsPanel>
-            <ItemsControl.ItemTemplate>
-                <DataTemplate>
-                    <Border BorderBrush="Black" BorderThickness="1" CornerRadius="5" Margin="2" Width="150" Height="300" Padding="10">
-                        <Border.Style>
-                            <Style TargetType="{x:Type Border}">
-                                <Setter Property="Background" Value="White"></Setter>
-                                <Style.Triggers>
-                                    <DataTrigger Binding="{Binding IsSelected}" Value="true">
-                                        <Setter Property="Background" Value="LightGray"></Setter>
-                                    </DataTrigger>
-                                </Style.Triggers>
-                            </Style>
-                        </Border.Style>
-                        <Grid>
-                            <Grid.RowDefinitions>
-                                <RowDefinition></RowDefinition>
-                                <RowDefinition></RowDefinition>
-                                <RowDefinition Height="Auto"></RowDefinition>
-                            </Grid.RowDefinitions>
-                            <UniformGrid Columns="2">
-                                <TextBlock Text="Name" />
-                                <TextBlock Text="{Binding Corporation.Name}"></TextBlock>
-                            </UniformGrid>
-                            <UniformGrid Grid.Row="1" Columns="2">
-                                <TextBlock Text="Credits" />
-                                <TextBlock Text="{Binding Corporation.StartingMoney}"></TextBlock>
-                            </UniformGrid>
-
-                            <ToggleButton Grid.Row="2" Content="Select" Command="{Binding RelativeSource={RelativeSource AncestorType={x:Type ItemsControl}}, Path=DataContext.SelectCardCommand}"
-                                          CommandParameter="{Binding}"
-                                          IsChecked="{Binding IsSelected, Mode=OneWay}" />
-                        </Grid>
-                    </Border>
-                </DataTemplate>
-            </ItemsControl.ItemTemplate>
-        </ItemsControl>
-        <ItemsControl Grid.Row="1" ItemsSource="{Binding PatentChoices}">
-            <ItemsControl.ItemsPanel>
-                <ItemsPanelTemplate>
-                    <StackPanel Orientation="Horizontal"></StackPanel>
-                </ItemsPanelTemplate>
-            </ItemsControl.ItemsPanel>
-            <!--<ItemsControl.ItemTemplate>
-                <DataTemplate>
-                    <Border BorderBrush="Black" BorderThickness="1" CornerRadius="5" Margin="2" Width="150" Height="300" Padding="10">
-                        <Border.Style>
-                            <Style TargetType="{x:Type Border}">
-                                <Setter Property="Background" Value="White"></Setter>
-                                <Style.Triggers>
-                                    <DataTrigger Binding="{Binding IsSelected}" Value="true">
-                                        <Setter Property="Background" Value="LightGray"></Setter>
-                                    </DataTrigger>
-                                </Style.Triggers>
-                            </Style>
-                        </Border.Style>
-                        <Grid>
-                            <Grid.RowDefinitions>
-                                <RowDefinition></RowDefinition>
-                                <RowDefinition></RowDefinition>
-                                <RowDefinition Height="Auto"></RowDefinition>
-                            </Grid.RowDefinitions>
-                            <UniformGrid Columns="2">
-                                <TextBlock Text="Name" />
-                                <TextBlock Text="{Binding Patent.Name}"></TextBlock>
-                            </UniformGrid>
-                            <UniformGrid Grid.Row="1" Columns="2">
-                                <TextBlock Text="Credits" />
-                                <TextBlock Text="{Binding Patent.BaseCost}"></TextBlock>
-                            </UniformGrid>
-
-                            <ToggleButton Grid.Row="2" Content="Select" Command="{Binding RelativeSource={RelativeSource AncestorType={x:Type ItemsControl}}, Path=DataContext.SelectCardCommand}"
-                                          CommandParameter="{Binding}"
-                                          IsChecked="{Binding IsSelected, Mode=OneWay}" />
-                        </Grid>
-                    </Border>
-                </DataTemplate>
-            </ItemsControl.ItemTemplate>-->
-        </ItemsControl>
-        <Button Grid.Row="2" Content="Validate" Command="{Binding CloseCommand}"></Button>
-    </Grid>
-</Window>
