--- TM.Digital.Client/NotifierBase.cs
@@ -1,44 +1,0 @@
-﻿using System;
-using System.ComponentModel;
-using System.Windows.Input;
-
-namespace TM.Digital.Client
-{
-    public class NotifierBase : INotifyPropertyChanged
-    {
-        public event PropertyChangedEventHandler PropertyChanged;
-
-        protected virtual void OnPropertyChanged(string e)
-        {
-            PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(e));
-        }
-    }
-
-    public class RelayCommand : ICommand
-    {
-        private readonly Action<object> _execute;
-        private readonly Func<object, bool> _canExecute;
-
-        public event EventHandler CanExecuteChanged
-        {
-            add { CommandManager.RequerySuggested += value; }
-            remove { CommandManager.RequerySuggested -= value; }
-        }
-
-        public RelayCommand(Action<object> execute, Func<object, bool> canExecute = null)
-        {
-            this._execute = execute;
-            this._canExecute = canExecute;
-        }
-
-        public bool CanExecute(object parameter)
-        {
-            return this._canExecute == null || this._canExecute(parameter);
-        }
-
-        public void Execute(object parameter)
-        {
-            this._execute(parameter);
-        }
-    }
-}
