+++ TM.Digital.Client/Resources/Common.xaml
