+++ TM.Digital.Client/Resources/Icons.xaml
@@ -1,7 +1,6 @@
 ﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-                    xmlns:local="clr-namespace:TM.Digital.Client"
-                    xmlns:cards="clr-namespace:TM.Digital.Model.Cards;assembly=TM.Digital.Model">
+                    xmlns:templateSelectors="clr-namespace:TM.Digital.Client.Resources.TemplateSelectors">
 
     <Path x:Key="leaf"  Data="M17,8C8,10 5.9,16.17 3.82,21.34L5.71,22L6.66,19.7C7.14,19.87 7.64,20 8,20C19,20 22,3 22,3C21,5 14,5.25 9,6.25C4,7.25 2,11.5 2,13.5C2,15.5 3.75,17.25 3.75,17.25C7,8 17,8 17,8Z" />
     <Path x:Key="starbox"  Data="M19,3A2,2 0 0,1 21,5V19C21,20.11 20.1,21 19,21H5A2,2 0 0,1 3,19V5A2,2 0 0,1 5,3H19M15.58,17L14.63,12.92L17.79,10.19L13.62,9.83L12,6L10.38,9.84L6.21,10.2L9.37,12.93L8.42,17L12,14.84L15.58,17Z" />
@@ -132,7 +131,7 @@
         </Grid>
     </DataTemplate>
 
-    <local:ResourceBoardTemplateSelector x:Key="resBoardSelector"
+    <templateSelectors:ResourceBoardTemplateSelector x:Key="resBoardSelector"
                                          Money="{StaticResource res_megacredits}"
                                          Steel="{StaticResource res_steel}"
                                          Titanium="{StaticResource res_titanium}"
@@ -225,7 +224,7 @@
         </Grid>
     </DataTemplate>
 
-    <local:TagTemplateSelector x:Key="tagTemplateSelector"
+    <templateSelectors:TagTemplateSelector x:Key="tagTemplateSelector"
                                Animal="{StaticResource dt_tag_animal}"
                                Plant="{StaticResource dt_tag_plant}"
                                Science="{StaticResource dt_tag_science}"
@@ -237,5 +236,5 @@
                                Earth="{StaticResource dt_tag_earth}"
                                Energy="{StaticResource dt_tag_energy}"
                                Event="{StaticResource dt_tag_event}">
-    </local:TagTemplateSelector>
+    </templateSelectors:TagTemplateSelector>
 </ResourceDictionary>
