+++ TM.Digital.Client/Resources/TagsCount.cs
@@ -1,0 +1,10 @@
+﻿using TM.Digital.Model.Cards;
+
+namespace TM.Digital.Client.Resources
+{
+    public class TagsCount
+    {
+        public Tags Tag { get; set; }
+        public int Count { get; set; }
+    }
+}
