+++ TM.Digital.Client/Resources/Templates/BoardPlacesTemplates.xaml
