+++ TM.Digital.Client/Resources/Templates/EffectTemplates.xaml
