+++ TM.Digital.Client/Resources/Templates/PlayerResources.xaml
@@ -2,9 +2,12 @@
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
                     xmlns:local="clr-namespace:TM.Digital.Client"        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
                     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+                    xmlns:main="clr-namespace:TM.Digital.Client.Screens.Main"
+                    xmlns:converters="clr-namespace:TM.Digital.Client.Resources.Converters"
+                    xmlns:resources="clr-namespace:TM.Digital.Client.Resources"
                     mc:Ignorable="d">
 
-    <DataTemplate DataType="{x:Type local:PlayerSelector}">
+    <DataTemplate DataType="{x:Type main:PlayerSelector}">
         <Grid>
             <Grid.ColumnDefinitions>
                 <ColumnDefinition Width="Auto"></ColumnDefinition>
@@ -13,7 +16,7 @@
 
             <Grid>
                 <Grid.Resources>
-                    <local:PlayedCardToTagsCountConverter x:Key="tagConverter" />
+                    <converters:PlayedCardToTagsCountConverter x:Key="tagConverter" />
                 </Grid.Resources>
                 <Grid.RowDefinitions>
                     <RowDefinition Height="Auto"></RowDefinition>
@@ -61,7 +64,7 @@
                 <ItemsControl Grid.Row="2" ItemsSource="{Binding Player.PlayedCards, Converter={StaticResource tagConverter}}" Margin="5">
                     <ItemsControl.ItemTemplate>
                         <DataTemplate>
-                            <Border d:DataContext="{d:DesignInstance {x:Type local:TagsCount}}" Padding="1" CornerRadius="2" BorderBrush="Black" BorderThickness="1" Margin="1">
+                            <Border d:DataContext="{d:DesignInstance {x:Type resources:TagsCount}}" Padding="1" CornerRadius="2" BorderBrush="Black" BorderThickness="1" Margin="1">
                                 <Grid  Margin="1">
                                     <Grid.ColumnDefinitions>
                                         <ColumnDefinition Width="Auto" />
