+++ TM.Digital.Client/Screens/HandSetup/GameSetupResources.xaml
@@ -1,0 +1,30 @@
+﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+                    xmlns:local="clr-namespace:TM.Digital.Client"
+                    xmlns:handSetup="clr-namespace:TM.Digital.Client.Screens.HandSetup">
+    <DataTemplate DataType="{x:Type handSetup:GameSetupViewModel}">
+        <Grid>
+            <Grid.RowDefinitions>
+                <RowDefinition></RowDefinition>
+                <RowDefinition></RowDefinition>
+                <RowDefinition Height="Auto"></RowDefinition>
+            </Grid.RowDefinitions>
+            <ItemsControl ItemsSource="{Binding CorporationChoices}" Visibility="{Binding IsInitialSetup, Converter={StaticResource boolToVisConverter}}">
+                <ItemsControl.ItemsPanel>
+                    <ItemsPanelTemplate>
+                        <StackPanel Orientation="Horizontal"></StackPanel>
+                    </ItemsPanelTemplate>
+                </ItemsControl.ItemsPanel>
+            </ItemsControl>
+            <ItemsControl Grid.Row="1" ItemsSource="{Binding PatentChoices}">
+                <ItemsControl.ItemsPanel>
+                    <ItemsPanelTemplate>
+                        <StackPanel Orientation="Horizontal"></StackPanel>
+                    </ItemsPanelTemplate>
+                </ItemsControl.ItemsPanel>
+
+            </ItemsControl>
+            <Button Grid.Row="2" Content="Validate" Command="{Binding CloseCommand}"></Button>
+        </Grid>
+    </DataTemplate>
+</ResourceDictionary>
