+++ TM.Digital.Client/Screens/HandSetup/SeyupCompletedEventHandler.cs
@@ -1,0 +1,4 @@
+﻿namespace TM.Digital.Client.Screens.HandSetup
+{
+    public delegate void SeyupCompletedEventHandler(GameSetupViewModel vm);
+}
