+++ TM.Digital.Client/Screens/Main/MainWindow.xaml
@@ -1,13 +1,14 @@
-﻿<Window x:Class="TM.Digital.Client.MainWindow"
+﻿<Window x:Class="TM.Digital.Client.Screens.Main.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:TM.Digital.Client"
+        xmlns:local="clr-namespace:TM.Digital.Client" 
+        xmlns:main="clr-namespace:TM.Digital.Client.Screens.Main"
         mc:Ignorable="d"
         WindowStartupLocation="CenterScreen"
         WindowState="Maximized"
-        d:DataContext="{d:DesignInstance {x:Type local:MainWindowViewModel}}"
+        d:DataContext="{d:DesignInstance {x:Type main:MainWindowViewModel}}"
         Title="MainWindow"  d:DesignHeight="800" d:DesignWidth="1000">
     <Grid x:Name="rootContent">
         <Grid.RowDefinitions>
@@ -62,5 +63,17 @@
             </ContentControl.Style>
         </ContentControl>
 
+        <ContentControl Content="{Binding GameSetupVm}" Grid.Row="0" Grid.RowSpan="3">
+            <ContentControl.Style>
+                <Style TargetType="{x:Type ContentControl}">
+                    <Setter Property="Visibility" Value="Collapsed" />
+                    <Style.Triggers>
+                        <DataTrigger Binding="{Binding GameSetupVm.IsVisible}" Value="true">
+                            <Setter Property="Visibility" Value="Visible" />
+                        </DataTrigger>
+                    </Style.Triggers>
+                </Style>
+            </ContentControl.Style>
+        </ContentControl>
     </Grid>
 </Window>
