+++ TM.Digital.Client/Screens/Menu/GameCreatedEventHandler.cs
@@ -1,0 +1,6 @@
+﻿using TM.Digital.Model.Game;
+
+namespace TM.Digital.Client.Screens.Menu
+{
+    public delegate void GameCreatedEventHandler(GameSessionInformation gameId);
+}
