+++ TM.Digital.Client/Screens/Menu/GameJoinSuccessEventHandler.cs
@@ -1,0 +1,6 @@
+﻿using TM.Digital.Model.Player;
+
+namespace TM.Digital.Client.Screens.Menu
+{
+    public delegate void GameJoinSuccessEventHandler(Player joindPlayer);
+}
