+++ TM.Digital.Client/Screens/Wait/WaitingGameScreenViewModel.cs
@@ -1,10 +1,11 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Collections.ObjectModel;
-using System.Text;
+using System.Windows;
+using TM.Digital.Client.Services;
+using TM.Digital.Client.ViewModelCore;
 using TM.Digital.Model.Game;
 
-namespace TM.Digital.Client
+namespace TM.Digital.Client.Screens.Wait
 {
     public class WaitingGameScreenViewModel : NotifierBase
     {
@@ -29,11 +30,20 @@ namespace TM.Digital.Client
             return IsOwner;
         }
 
-        private void ExecuteStartGame(object obj)
+        private async void ExecuteStartGame(object obj)
         {
-            IsVisible = false;
-            IncommingMessages.Clear();
-            InitialMessage = string.Empty;
+            if (await _apiCaller.StartGame(Session.GameSessionId))
+            {
+                IsVisible = false;
+                IncommingMessages.Clear();
+                InitialMessage = string.Empty;
+            }
+            else
+            {
+                MessageBox.Show("Can't start game...");
+            }
+            
+            
         }
 
         public RelayCommand StartGameCommand { get; set; }
