+++ TM.Digital.Client/ViewModelCore/NotifierBase.cs
@@ -1,0 +1,14 @@
+﻿using System.ComponentModel;
+
+namespace TM.Digital.Client.ViewModelCore
+{
+    public class NotifierBase : INotifyPropertyChanged
+    {
+        public event PropertyChangedEventHandler PropertyChanged;
+
+        protected virtual void OnPropertyChanged(string e)
+        {
+            PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(e));
+        }
+    }
+}
