+++ TM.Digital.Client/ViewModelCore/RelayCommand.cs
@@ -1,0 +1,33 @@
+﻿using System;
+using System.Windows.Input;
+
+namespace TM.Digital.Client.ViewModelCore
+{
+    public class RelayCommand : ICommand
+    {
+        private readonly Action<object> _execute;
+        private readonly Func<object, bool> _canExecute;
+
+        public event EventHandler CanExecuteChanged
+        {
+            add { CommandManager.RequerySuggested += value; }
+            remove { CommandManager.RequerySuggested -= value; }
+        }
+
+        public RelayCommand(Action<object> execute, Func<object, bool> canExecute = null)
+        {
+            this._execute = execute;
+            this._canExecute = canExecute;
+        }
+
+        public bool CanExecute(object parameter)
+        {
+            return this._canExecute == null || this._canExecute(parameter);
+        }
+
+        public void Execute(object parameter)
+        {
+            this._execute(parameter);
+        }
+    }
+}
