+++ TM.Digital.Model/ServerPushMethods.cs
@@ -1,13 +1,10 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace TM.Digital.Model
+﻿namespace TM.Digital.Model
 {
-   public static class ServerPushMethods
-   {
-       public const string RecieveGameUpdate = "ReceiveGameUpdate";
-       public const string PlaceTileRequest = "PlaceTile";
-       public const string PlayerJoined = "PlayerJoined";
-   }
+    public static class ServerPushMethods
+    {
+        public const string SetupChoice = "SetupPlayerChoice";
+        public const string RecieveGameUpdate = "ReceiveGameUpdate";
+        public const string PlaceTileRequest = "PlaceTile";
+        public const string PlayerJoined = "PlayerJoined";
+    }
 }
